+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Update latest dependencies, Add .gitattributes file
 
+## [Unreleased]
+
+### Added
+
+- Add is-main-branch command to check for main branch in Git
+
 ## [3.1.0] - 2026-03-18
 
 ### Changed
@@ -194,6 +200,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_publish/compare/3.1.0...HEAD
 [3.1.0]: https://github.com/inlavigo/gg_publish/compare/3.0.18...3.1.0
 [3.0.18]: https://github.com/inlavigo/gg_publish/compare/3.0.17...3.0.18
 [3.0.17]: https://github.com/inlavigo/gg_publish/compare/3.0.16...3.0.17